third_party/WebKit/LayoutTests/fast/css/getComputedStyle/getComputedStyle-border-width-expected.txt

Tests that the border-width shorthand is computed properly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS computedStyle.getPropertyValue('border-width') is '10px 5px 4px 3px'
PASS computedStyle.getPropertyValue('border-width') is '320px 160px 64px 80px'
PASS computedStyle.getPropertyValue('border-width') is '10px 0px 0px'
PASS computedStyle.getPropertyValue('border-width') is '0px'
PASS computedStyle.getPropertyValue('border-width') is '0px'
PASS successfullyParsed is true

TEST COMPLETE
